fix mac build

+++ vowpalwabbit/array_parameters.h
@@ -137,7 +137,7 @@ public:
 	      T(iter);
 	  }
 	
-	template<void(*T)(iterator&, size_t)> //for random initialization of weights (with stride) 
+	template<void(*T)(iterator&, uint64_t)> //for random initialization of weights (with stride) 
 	inline void set_default()
 	{  uint32_t stride = 1 << _stride_shift;
 	   iterator iter = begin();
@@ -145,7 +145,7 @@ public:
 			T(iter, i);
 	}
 
-	template<void(*T)(iterator&, size_t, uint32_t)> //for random initialization of the entire weight_vector 
+	template<void(*T)(iterator&, uint64_t, uint32_t)> //for random initialization of the entire weight_vector 
 	inline void set_default()
 	{ uint32_t stride = 1 << _stride_shift;
 	iterator iter = begin();
